+++ service-feature-table-nocache/src/main/java/com/esri/arcgisruntime/samples/servicefeaturetablenocache/MainActivity.java
@@ -18,7 +18,7 @@ package com.esri.arcgisruntime.samples.servicefeaturetablenocache;
 import android.os.Bundle;
 import android.support.v7.app.AppCompatActivity;
 
-import com.esri.arcgisruntime.datasource.arcgis.ServiceFeatureTable;
+import com.esri.arcgisruntime.data.ServiceFeatureTable;
 import com.esri.arcgisruntime.geometry.Envelope;
 import com.esri.arcgisruntime.geometry.SpatialReferences;
 import com.esri.arcgisruntime.layers.FeatureLayer;
